+++ README.md
@@ -32,11 +32,13 @@
 ### 3단계 요구사항
 
 - [x] 날짜와 테마를 기반으로 예약 가능한 시각과 예약 불가능한 시각을 응답할 수 있다.
-- [x] `/`요청 시 인기 테마 페이지를 응답한다. 
+- [x] `/`요청 시 인기 테마 페이지를 응답한다.
 - [ ] 지난 7일간의 기록으로 인기테마 목록 10개를 불러올 수 있다.
 
 ### 4단계 요구사항
+
 - [x] 로그인 페이지요청에 응답이 가능하다 (GET "/login")
 - [ ] 로그인 API 구현. (POST "login")
-  - [ ] 로그인 응답 cookie에 token값을 포함 시킨다.
+    - [ ] email, password로 User를 불러올 수 있다.
+    - [ ] 로그인 응답 cookie에 이름 기반의 token값을 포함 시킨다.
 - [ ] 사용자 정보 조회 API 구현 (GET "/login/check")
