+++ src/main/java/roomescape/domain/login/domain/User.java
@@ -1,16 +1,51 @@
 package roomescape.domain.login.domain;
 
+import java.util.Objects;
+
 public class User {
     private Long id;
-
+    private final String name;
     private final String email;
     private final String password;
-    private final String name;
 
-    public User(Long id, String email, String password, String name) {
+    public User(Long id, String name, String email, String password) {
         this.id = id;
+        this.name = name;
         this.email = email;
         this.password = password;
-        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        User user = (User) o;
+        return Objects.equals(id, user.id) && Objects.equals(name, user.name)
+                && Objects.equals(email, user.email) && Objects.equals(password, user.password);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name, email, password);
     }
 }
