+++ src/main/java/roomescape/domain/login/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package roomescape.domain.login.repository;
+
+import java.util.Optional;
+import roomescape.domain.login.domain.User;
+
+
+public interface UserRepository {
+
+    User insert(User user);
+
+    Optional<User> findById(Long id);
+}
